Update data generation and probe handling
- Add use_probe option to control probe application:
  * When enabled, the model uses probe data from the dataset
  * When disabled, it falls back to a generic, uninformative probe
- Modify MAE_ViT model to incorporate probe data in diffraction simulation
- Refactor mae_diff.py for improved readability and error handling
- fix visualization for illuminated object data in cat_images function
- Adjust diffraction_from_channels to return illuminated complex object

After these changes, the out of distribution loss is much worse than
before. This could be because of a silent change in the data's
normalization I switched from png to npz format. Another possibility:
using the specific (in distribution) probe is worse for out of
distribution generalization than using the generic, uninformative one